% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{clusters}
\alias{clusters}
\title{Returns clusters for n random countries}
\usage{
clusters(n = 10, centers = 3)
}
\arguments{
\item{n}{An integer of the number of countries}

\item{centers}{An integer of the number of clusters}
}
\value{
A plot
}
\description{
Returns clusters for n random countries
}
