% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_double}
\alias{extract_double}
\title{Extract double from xml data for a year}
\usage{
extract_double(year_xml, node)
}
\arguments{
\item{year_xml}{xml data for a year}

\item{node}{name of the child node}
}
\value{
A double
}
\description{
Extract double from xml data for a year
}
