% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_integer}
\alias{extract_integer}
\title{Extract integer from xml data for a year}
\usage{
extract_integer(year_xml, node)
}
\arguments{
\item{year_xml}{xml data for a year}

\item{node}{name of the child node}
}
\value{
An integer
}
\description{
Extract integer from xml data for a year
}
