% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic.R
\name{gdp}
\alias{gdp}
\title{Return GDP data of country}
\usage{
gdp(country_name)
}
\arguments{
\item{country_name}{The name of a country}
}
\value{
A data frame of gdp and year
}
\description{
Return GDP data of country
}
