% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic.R
\name{gdp_by_code}
\alias{gdp_by_code}
\title{Return GDP data of country}
\usage{
gdp_by_code(country_code)
}
\arguments{
\item{country_code}{The code of a country}
}
\value{
A data frame of gdp and year
}
\description{
Return GDP data of country
}
