% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic.R
\name{gdp_per_capita}
\alias{gdp_per_capita}
\title{Return GDP per capita data of country}
\usage{
gdp_per_capita(country_name)
}
\arguments{
\item{country_name}{The name of a country}
}
\value{
A data frame of gdp per capita and year
}
\description{
Return GDP per capita data of country
}
