% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.R
\name{parse_date}
\alias{parse_date}
\title{Parses date from xml data for a year}
\usage{
parse_date(year_xml)
}
\arguments{
\item{year_xml}{xml data for 1 year}
}
\value{
An integer year
}
\description{
Parses date from xml data for a year
}
