% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.R
\name{parse_population}
\alias{parse_population}
\title{Parses population from xml data for a year}
\usage{
parse_population(year_xml)
}
\arguments{
\item{year_xml}{xml data for 1 year}
}
\value{
An integer population
}
\description{
Parses population from xml data for a year
}
