% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.R
\name{population}
\alias{population}
\title{Return population data of country}
\usage{
population(country_name)
}
\arguments{
\item{country_name}{The name of a country}
}
\value{
A data frame of population and year
}
\description{
Return population data of country
}
