% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.R
\name{population_by_code}
\alias{population_by_code}
\title{Return population data of country by code}
\usage{
population_by_code(country_code)
}
\arguments{
\item{country_code}{The code of a country}
}
\value{
A data frame of population and year
}
\description{
Return population data of country by code
}
