% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{random_countries}
\alias{random_countries}
\title{Return random list of countries}
\usage{
random_countries(n)
}
\arguments{
\item{n}{An integer}
}
\value{
A vector
}
\description{
Return random list of countries
}
