% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{random_countries_data}
\alias{random_countries_data}
\title{Return data for n random countries}
\usage{
random_countries_data(n)
}
\arguments{
\item{n}{An integer}
}
\value{
A data frame with data for n countries
}
\description{
Return data for n random countries
}
